arm64のウォーニングに対応。

+++ Classes/ScLog.m
@@ -23,7 +23,7 @@ void ScMemoryReport()
                                    (task_info_t)&info,
                                    &size);
     if( kerr == KERN_SUCCESS ) {
-        NSLog(@"Memory in use (in MB): %u", info.resident_size / 1024 / 1024);
+        NSLog(@"Memory in use (in MB): %lu", (unsigned long)info.resident_size / 1024 / 1024);
     } else {
         NSLog(@"Error with task_info(): %s", mach_error_string(kerr));
     }
